+++ src/components/Home/index.js
@@ -40,6 +40,8 @@ const UserList = ({ users }) =>
     {Object.keys(users).map(key =>
       <div key={key}>{users[key].username}</div>
     )}
+
+    
   </div>
 
 const authCondition = (authUser) => !!authUser;
