+++ src/components/Navigation/index.js
@@ -18,6 +18,7 @@ const NavigationAuth = () =>
     <li><Link to={routes.LANDING}>Landing</Link></li>
     <li><Link to={routes.HOME}>Home</Link></li>
     <li><Link to={routes.ACCOUNT}>Account</Link></li>
+    <li><Link to={routes.LIST}>itemList</Link></li>
     <li><SignOutButton /></li>
   </ul>
