+++ src/constants/routes.js
@@ -4,3 +4,4 @@ export const SIGN_IN = '/signin';
 export const PASSWORD_FORGET = '/pw-forget';
 export const HOME = '/home';
 export const ACCOUNT = '/account';
+export const LIST = '/ListPage'
